fix: avoid player getting stuck on a teleport
The fix I used is to have the TP Ball keep a record of the last location it was in before each physics cycle. Now, the player will, instead of teleporting into its current position (which when the bug happened was partially inside a wall), it will teleport to this previous "safe" position.